Fix various tests and layout
