+++ xopt/tests/generators/sequential/test_extremum_seeking.py
@@ -3,6 +3,7 @@ import pytest
 from pydantic import ValidationError
 
 from xopt import Xopt
+from xopt.errors import SeqGeneratorError
 from xopt.evaluator import Evaluator
 from xopt.generators.sequential.extremumseeking import ExtremumSeekingGenerator
 from xopt.resources.testing import TEST_VOCS_BASE
@@ -14,7 +15,7 @@ class TestExtremumSeekingGenerator:
         gen = ExtremumSeekingGenerator(vocs=TEST_VOCS_BASE)
 
         # Try to generate multiple samples
-        with pytest.raises(ValueError):
+        with pytest.raises(SeqGeneratorError):
             gen.generate(2)
 
     def test_es_options(self):
@@ -259,3 +260,23 @@ class TestExtremumSeekingGenerator:
         assert np.allclose(cES, X.data["f"].to_numpy()[:-1]), (
             "Xopt ES does not match the vanilla one"
         )
+
+    def test_es_convergence(self):
+        def eval_f(x):
+            return {"y1": np.sum([x**2 for x in x.values()])}
+
+        variables = {f"x{i}": [-5, 5] for i in range(10)}
+        objectives = {"y1": "MINIMIZE"}
+        vocs = VOCS(variables=variables, objectives=objectives)
+        evaluator = Evaluator(function=eval_f)
+        generator = ExtremumSeekingGenerator(vocs=vocs)
+        X = Xopt(vocs=vocs, evaluator=evaluator, generator=generator)
+
+        X.random_evaluate(1)
+        for i in range(5000):
+            X.step()
+
+        idx, best, _ = X.vocs.select_best(X.data)
+        xbest = X.vocs.variable_data(X.data.loc[idx, :]).to_numpy().flatten()
+        assert best[0] >= 0.0
+        assert best[0] <= 0.001
