+++ xopt/tests/generators/sequential/test_rcds.py
@@ -2,7 +2,8 @@ import numpy as np
 import pytest
 from pydantic import ValidationError
 
-from xopt import Xopt
+from xopt import VOCS, Xopt
+from xopt.errors import SeqGeneratorError
 from xopt.generators.sequential.rcds import RCDSGenerator
 from xopt.resources.testing import TEST_VOCS_BASE
 
@@ -23,7 +24,7 @@ class TestRCDSGenerator:
         gen = RCDSGenerator(vocs=TEST_VOCS_BASE)
 
         # Try to generate multiple samples
-        with pytest.raises(ValueError):
+        with pytest.raises(SeqGeneratorError):
             gen.generate(2)
 
     def test_rcds_options(self):
@@ -65,3 +66,29 @@ class TestRCDSGenerator:
 
         X.generator.reset()
         assert not X.generator.is_active
+
+    def test_rcds_convergence(self):
+        def eval_f(x):
+            return {"y1": np.sum([x**2 for x in x.values()])}
+
+        variables = {f"x{i}": [-5, 5] for i in range(10)}
+        objectives = {"y1": "MINIMIZE"}
+        vocs = VOCS(variables=variables, objectives=objectives)
+
+        config = {
+            "generator": {
+                "name": "rcds",
+            },
+            "evaluator": {"function": eval_f},
+            "vocs": vocs,
+        }
+        X = Xopt.from_dict(config)
+        X.random_evaluate(1)
+        for i in range(1000):
+            X.step()
+        print(X.data["y1"])
+
+        idx, best, _ = X.vocs.select_best(X.data)
+        xbest = X.vocs.variable_data(X.data.loc[idx, :]).to_numpy().flatten()
+        assert best[0] >= 0.0
+        assert best[0] <= 0.001
